Feature/230 new moving layer inspector from left panel to bottom middle mobile (#241)
* Icons to react components

* update window size

* minimized layer inspector for mobi

* all functionality in.

* suggested improvements
