+++ src/components/play/PlayRoute.js
@@ -3,8 +3,8 @@ import PlayUI from './PlayUI'
 //import PatternsSidebar from './PatternsSidebar'
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/styles';
-import EffectsSidebar from './EffectsSidebar';
 import Box from '@material-ui/core/Box';
+//import EffectsSidebar from './EffectsSidebar';
 import _ from 'lodash';
 import Loader from 'react-loader-spinner';
 import { connect } from "react-redux";
@@ -15,8 +15,8 @@ import Instruments from '../../audio-engine/Instruments'
 import FX from '../../audio-engine/FX'
 import ShareDialog from '../dialogs/ShareDialog'
 import { getDefaultUserBus, getDefaultUserPatterns } from '../../utils/defaultData'
-import LayerSettingsAlt from './layer-settings/LayerSettingsAlt';
-import OrientationDialog from '../dialogs/OrientationDialog';
+import LayerSettings from './layer-settings/LayerSettings';
+//import OrientationDialog from '../dialogs/OrientationDialog';
 import CustomSamples from '../../audio-engine/CustomSamples';
 
 const styles = theme => ({
@@ -346,10 +346,13 @@ class PlayRoute extends Component {
                         visible={true}
                     />
                 }
-                <EffectsSidebar />
+                {/* <PatternsSidebar />
+                <EffectsSidebar /> */}
                 <ShareDialog />
-                <LayerSettingsAlt playUIRef={this.playUIRef} />
-                <OrientationDialog />
+                <Box style={{ position: 'relative', display: 'flex', justifyContent: 'center', width: '100%' }}>
+                    <LayerSettings playUIRef={this.playUIRef} />
+                </Box>
+                {/* <OrientationDialog /> */}
             </Box>
         )
     }
