+++ src/components/play/layer-settings/resources/index.js
@@ -11,4 +11,3 @@ export { default as MutedIcon } from './Muted'
 export { default as ErasorIcon } from './Erasor'
 export { default as TrashIcon } from './Trash'
 export { default as ElipsisIcon } from './Elipsis'
-export { default as BackButton } from './BackButton'
